Create Shanhai Jing – The depths of the cloud
This is a 2D side-scrolling two-player co-operative adventure game. The two protagonists meet through a timeline centuries apart and join forces to seal the four divine beasts. In the game players use time differences with each other and the scrolls to decipher the game and cooperate with each other to defeat the monsters in the Shan Hai Jing.